test on windows

+++ includes/loc.h
@@ -2,11 +2,12 @@
 #define LIB_RUBY_PARSER_LOC_H
 
 #include <stddef.h>
+#include <stdint.h>
 
 struct Loc
 {
-    size_t begin;
-    size_t end;
+    uint32_t begin;
+    uint32_t end;
 };
 
 #endif // LIB_RUBY_PARSER_LOC_H
